reorganize for packaging and distribution
