+++ cambrinary/__init__.py
@@ -1,0 +1,3 @@
+"""A Linux terminal online dictionary, based on `Cambridge Dictionary <https://dictionary.cambridge.org>`_"""
+
+__version__ = '0.0.1'
