+++ cambrinary/requirements.in
@@ -1,0 +1,2 @@
+aiohttp
+beautifulsoup4
