+++ cambrinary/requirements.txt
@@ -1,0 +1,9 @@
+aiohttp==3.6.1
+async-timeout==3.0.1      # via aiohttp
+attrs==19.1.0             # via aiohttp
+beautifulsoup4==4.8.0
+chardet==3.0.4            # via aiohttp
+idna==2.8                 # via yarl
+multidict==4.5.2          # via aiohttp, yarl
+soupsieve==1.9.3          # via beautifulsoup4
+yarl==1.3.0               # via aiohttp
