--- requirements.txt
@@ -1,2 +1,0 @@
-aiohttp==3.5.4
-beautifulsoup4==4.7.1
